MR-57: Wrong urls is crashing UrlTokenizer
Changed the check from [if(url.Length < 2)] to [if(parts.Length < 2)]. This will still throw an exception however it will be UrlTokenizerException("Invalid url") instead of IndexOutOfRangeException. I have also added a test for a bad url. Please change the test to demand the desired behavior and I or someone else can implement.

dru

git-svn-id: http://svn.castleproject.org:8080/svn/castle/trunk@1279 73e77b4c-caa6-f847-a29a-24ab75ae54b6