- Fixed DYNPROXY-ISSUE-96 - Implement order of precenence for interfaces on proxy. This is a pretty major change (quite a lot of code was changed, and some parts were fundamentally altered) so I'l appreciate if someone decided to dedicate some time to reviewing these changes.
It introduces some significant changes of behavior (take a look at the description of the issue). Also as a side-effect when an interface method is implemented virtually, it will now generate two members on the proxy - one overriding virtual method on the class, second implementing interface method, both forwarding to the base implementation. This is probably not a valuable thing to have, so I'm gonna remove that in a future commit.

I also Ignored one test, that would fail because of change in the architecture. This is temporary, and fixing DYNPROXY-ISSUE-94 will make that test pass again.The new bahavior that made the test fail has no impact on stability, so it's not a big deal.

git-svn-id: http://svn.castleproject.org:8080/svn/castle/trunk@5894 73e77b4c-caa6-f847-a29a-24ab75ae54b6