Update log4net reference in a bad way
As Krzysztof noted in my first pull request (and what prompted this
whole exercise in the first place), log4net changed their assembly
signing key in the new version so bindingRedirects in *.config files
don't work.  This approach creates a duplicate project, which builds
Castle.Core.log4netintegration-2 which is linked to the new log4net.
With this (and another NuGet package) consumers can choose which log4net
integration to use so it doesn't hold back all of their other updates.