Added SourceContext to the Serilog Logger
Without this there is no way to capture the name / context (commonly the type name) provided by a facility while wiring a logger.

AFAIK .ForContext("SourceContext", name); Has the same behavior as .ForContext<T>() in Serilog.

If someone has a better idea for the string constant? Maybe a constructor parameter?