Make `ProxyUtil.IsAccessible` check declaring type
This commit fixes the bug revealed in the preceding commit by making
`ProxyUtil.IsAccessible(MethodBase)` take into account the declaring
type's accessibility, as well.

That method is currently a forward to the identically named method in
`InternalsUtils` which also receives internals calls from DynamicProxy
itself, so we need to clearly separate the public API and the internal
API before we change any behaviour. This is done as follows:

 * The relationship between `ProxyUtil` and `InternalsUtil` is revers-
   ed, i.e. the actual implementations are moved to `ProxyUtil`, and
   `InternalsUtil`'s public methods are reduced to forwards.

 * In `ProxyUtil`, we now have internal methods (e.g. `IsAccessible-
   Method` and `IsAccessibleType`) which get called by DynamicProxy.
   We also have a separate public method group (`IsAccessible`) which
   makes use of the internal methods.

Finally, we move in suitable preexisting code from `DefaultProxyBuil-
der` that allows checking a declaring type's visibility.