Add .NET Standard 1.5 target
The `in` modifier on parameters that was introduced with C# 7.2 gets
encoded as a modreq. Because .NET Standard 1.3 (the sole `netstandard`
TFM currently targeted by Castle Core) does not support emitting
custom modifiers in IL, we couldn't support proxying methods with such
an `in` modifier.

However we can support this scenario by adding `netstandard1.5` as a
TFM to Castle Core: .NET Standard 1.5 supports emitting modreqs in IL.