Add a new, non-strong-named unit test project
Prior to this commit, DynamicProxy, when asked to proxy a type from a
weak-named assembly, would generate the proxy type in the weak-named
dynamic assembly, and the associated invocation types in the strong-
named dynamic assembly. This results in PEVerify rightfully claiming
that both of them are invalid.

Since the assembly of main test project (Castle.Core.Tests) gets
signed, it's difficult to write tests regarding this issue. Therefore
let's add a second test project that stays non-strong-named.