Configure NuGet package sources at project level
In order to compile for the .NET Framework on Unix, we need suitable
reference assemblies. Those are brought in via the unofficial NuGet
package `Microsoft.TargetingPack.NETFramework.v4.6.1`, which is not
available on nuget.org, but on a different MyGet feed.

The latest stable version of Mono to which we've just migrated (see
previous commit) now uses Microsoft's tools (e.g. msbuild instead of
xbuild) and this means that NuGet package restoration on Travis CI
now happens before `build.sh` starts. Therefore we need to reconfigure
NuGet so it finds the reference assembly package by itself.