Remove MyGet & FrameworkPathOverride from projects
It appears that Mono 5.10.1 comes with the .NET 4.6.1 reference assem-
blies that we need to build for the `net461` target framework. So we
can use those instead of fetching them from some MyGet package.

Also, `<FrameworkPathOverride>` can be moved out of all project files
and be replaced by an environment variable in the build script.