Replace internal use of `Lock` in `ModuleScope`
Unfortunately, `ModuleScope` exposes a `Lock` for its proxy type cache
publicly. This means we cannot just remove its `Lock` completely, but
we can start replacing it with a `ReaderWriterLockSlim` internally.

A new `Lock.CreateFor(ReaderWriterLockSlim)` method is needed so that
a `Lock` can be exposed that is mapped to the `ReaderWriterLockSlim`
that is used internally.