Abstract away `ReaderWriterLockSlim` in most cases
`ReaderWriterLockSlim` is used mostly in conjunction with regular
dictionaries. The double-checked lock pattern is used in many places
to query and update those dictionaries.

Instead of repeating this same pattern over and over again, abstract
it away with a dictionary-like type `SynchronizedDictionary<,>` that
performs the required locking pattern.

`ModuleScope`'s internal methods for accessing the type cache can now
be replaced with a simple internal property `TypeCache`.