Reinstate upgradeable read locks
Based on a review comment, this reinstates the upgradeable read locks
that were originally taken before entering write locks.

The intent of this PR is to deprecate `Lock`, not to make functional
changes. While forgoing the upgradeable read locks in favor of enter-
ing write locks right away might *possibly* be an optimization, this
is a different concern and would need to be properly established using
careful benchmarking. Let's do this another time.