Remove redundant call to EmitCallEnsureValidTarget (#397)
* Remove redundant call to EmitCallEnsureValidTarget

`EnsureValidTarget` exists only on `CompositionInvocation`, but not
on `InheritanceInvocation` (the only other invocation type). Emitting
a call to it makes sense *only* for the former invocation type. Since
its generator type (`CompositionInvocationTypeGenerator`) already un-
conditionally emits a call to `EnsureValidTarget`, doing so again in
the base class `InvocationTypeGenerator` is redundant.

* Rename two `InvocationMethods` members