Deprecate `ConstReference`
`ConstReference` is misdesigned: on the one hand, it is used to repre-
sent constant values, and on the other, it claims to be a `Reference`
(that is, a location whose address can be taken and in which one can
store values).

`ConstReference` should really be an `Expression`. Because it inherits
from `Reference` instead, it ends up implementing most of its methods
by throwing exceptions, and all uses must be combined with a call to
the `.ToExpression()` conversion method.

Deprecate `ConstReference` by introducing more efficient types belong-
ing to the same family as `LiteralIntExpression`, then replace all
uses of `ConstReference` with the new types.