Deprecate `StatementExpression`
`MethodInvocationExpression` currently inherits from `Expression`,
but it can also be used as a `Statement`. In those cases, `Statement-
Expression` is used as an adapter for that purpose. It is however
completely superfluous, because in order to add a `MethodInvocation-
Expression` to an `AbstractCodeBuilder`, we can simply use its `Add-
Expression` method instead of `AddStatement`, thus making the adapter
type redundant.

(Ideally, `Expression` and `Statement` should have been interfaces
and `MethodInvocationExpression` would have implemented them both.)