Skip failing tests (runtime bug)
It turns out that the tests from the previous commit fail due to a bug
in the runtime (`in` parameters + generic types cause `ldtoken` IL to
reference a non-existent method) that we've previously encountered.

Skip these tests and replace them with similar ones that do not rely
on a generic delegate type. These will pass just fine.