Add `IInvocation.GetProceedInfo()` method
which returns a descriptor for whatever a call to `Proceed` would pro-
ceed to. For now, this can only be used to capture the `Proceed` call
for later execution (e.g. from an async continuation).

More reflection capabilities could be added later on (e.g. for discov-
ering whether `Proceed` would invoke another interceptor or the proxy
target object).