Accuracy of docs & exception regarding `Proceed`
ea8fa099 made it permissible to call `invocation.Proceed()` more than
once from the same interceptor. The introductory article in the docs
does not yet reflect that.

That change also removed an error case for which `AbstractInvocation`
goes to great lengths to build an error message. Today it is no longer
needed as the error will only get triggered by seriously faulty client
code. Replace it with a simpler constant message.

Finally, this slightly increases the accuracy of the description of
how the interception pipeline and `Proceed` work.