Make non-proxied methods visible to generator
Methods which are filtered out by `MembersCollector` (e.g. because the
hook's `ShouldInterceptMethod` returned false) are hereby made visible
to the proxy-class-with-target generator.

Some methods must still be excluded: `ForwardingMethodGenerator` does
not perform any accessibility checks, so we need to check whether the
target method is callable "from the outside". For most cases, a simple
`method.IsPublic` check suffices, however we might want to test if the
target method is callable through an interface.