Remove <RuntimeIdentifier> from test projects
This was apparently used as a hack to force the `dotnet` CLI to create
an EXE file without actually publishing the project.

Unfortunately, it now causes the build to fail with a NETSDK1061 error
because it leads to different versions of the .NET SDK being used for
package restore & build.

Let's instead just use `dotnet` to run the test projects' DLLs, as we
are already doing in the Linux build script.