Make NUnit tests runnable in VS Test Explorer
While the test projects can simply be started to run the tests (thanks
to NUnitLite), tests cannot be run in VS Test Explorer, which could be
handy especially when one wants to run only a subset of all tests.
Currently, doing that requires one to set a `--test=...` command-line
argument in the test project settings.)

This commit adds the packages required to run our tests in VS Test
Explorer.