Merge `.travis.yml` into AppVeyor build matrix
This is a fairly straight-forward rewrite of `.appveyor.yml`. A second
image `Ubuntu` gets added, and the previous contents are mostly made
specific to the preexisting `Visual Studio 2017` image.

Because the Windows build does the major part of the work, including
test results upload & deployment, we put it in the last position. This
isn't strictly necessary, as the `on_success` script should only run
when all build jobs succeeded; but let's stay on the safe side.