Add failing test re: method name collision
When deciding how to name interface methods on the proxy type, Dynamic-
Proxy attempts to prevent name collisions by switching to explicit
implementation (where method names get prefixed with the declaring
type's name).

As this test shows, this strategy is not sufficient to prevent name
collisions, because the namespace does not get included in type names.