Remove security permissions in `DefaultSmtpSender`
The use of security permissions' and `.IsGranted()` here appears to
have had the purpose of allowing this class to work correctly when no
credentials are set and executing code would have no permission to
read `NetworkCredentials`.

We can get rid of security permissions by refactoring code such that
the `NetworkCredential` gets created only on demand. The only observ-
able difference should be that in the absence of certain granted perm-
issions, querying or setting `UserName`, `Password`, or `Domain` will
no longer throw immediately. Failure is deferred until the moment when
the SMTP client is built.