Don't pass around `interfaces` where not required
Also, standardize on `allInterfaces` as the name for the return value of
`GetTypeImplementerMapping` to distinguish it better from `interfaces`.

(This is better than the other previous variant--`implementedInterfaces`
when they aren't implemented just yet--but still not completely accurate
as it may also include delegate types. Maybe those names will be "fixed"
someday.)