DynamicProxy puts delegate types in wrong module
Back in 95383909 an issue was fixed where DynamicProxy would not always
put a generated invocation type in the same dynamic module as the proxy
type relying on it. It appears the same issue still exists for generated
delegate types. (Those are generated for class proxies when proceeding
from intercepted protected methods to the target.)