Fix interception of inherited default impls in class proxies
A class implementing an interface with default method implementations
will not necessarily override those methods; in such cases, the methods
are currently "invisible" to DynamicProxy and thus won't get proxied.

This can be solved by an extra pass over all interfaces to collect such
methods, as is done in this commit.

Ideally we wouldn't have to iterate over all interfaces more than once,
so we should next make an effort to limit the runtime impact of this
extra pass.