Refactor "special case for handling arrays"
This is where the `NullReferenceException` happens.

The pre-existing comment here didn't explain *why* this special case is
needed at all, which only became apparent when I temporarily removed it.

Let's explain the intent behind the special case better, and move it
inside an `if` that triggers only when the right conditions are present
(which now excludes `null` values).